requirements.txt
certifi==2024.7.4
fastapi==0.115.0
langchain==0.3.1
langchain_chroma==0.1.4
langchain_community==0.3.1
langchain_core==0.3.8
langchain_experimental==0.3.2
langchain_huggingface==0.1.0
langchain_mongodb==0.2.0
motor==3.6.0
passlib==1.7.4
pydantic==2.9.2
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1